like review into component
